mock/requirements/base.txt
aiohttp==3.8.1
asyncio==3.4.3
cuid==0.3
fastapi==0.74.0
mypy==0.931
pydantic-factories==1.2.6
pytz==2021.3
requests==2.27.1
sentry-sdk==1.5.5
uvicorn==0.17.5
